Moved 2 Role-specific Events to their corresponding sub-package
